refactor: 시간값 유효성 확인 로직 서비스 레이어로 이동
